[feature/asset-domain]fix: Asset 잘못된 표기 수정

+++ lms/src/main/java/com/lms/domain/asset/rebuild/RebuildAsset.java
@@ -3,11 +3,11 @@ package com.lms.domain.asset.rebuild;
 import com.lms.domain.asset.UploadStatus;
 
 public record RebuildAsset(
-    Integer id,
+    Long id,
     String mimeType,
     String path,
     String originalFilename,
     String convertedFilename,
-    Integer contentId,
+    Long contentId,
     UploadStatus status
 ) {}
